Prevent symbols not being exported in Qt5WebKit.dll
WebKit svn commit 136242 implemented a split into QtWebKit and QtWebKitWidgets.
Due to this change a static library named WebKit1.a is created first.
After this a shared library is created named Qt5WebKit.dll which contains
the contents of this static library and some other object files.
However, various symbols in the static library are expected to be exported
in the Qt5WebKit.dll shared library. As static libraries normally don't care
about exported symbols (everything is exported after all) the decoration
attribute Q_DECL_EXPORT won't be set.
This results in undefined references when trying to link the QtWebKitWidgets
shared library (which depends on various symbols which aren't exported properly
in the Qt5WebKit.dll shared library).